Specify a pre-release version (0.13.0) for now.
Hashicorp merged a pull request on June 19th that prevents GitHub
actions without a pinned version from using pre-release versions
(i.e. 0.13.0 as of this writing):

https://github.com/hashicorp/setup-terraform/pull/19

However, this run appears to have used 0.13.0 before it was merged:

https://github.com/CovidShield/server/runs/786126855

I think we can still use the prerelease version if we specify it
explicitly.